re-work of axios get requests as well as useEffects for DRYer code
