+++ src/components/ServicesContainer.jsx
@@ -1,6 +1,5 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import Axios from 'axios';
 import Service from './Service/Service';
 import './ServicesContainer.css'
 
@@ -9,14 +8,6 @@ const backendUrl = process.env.REACT_APP_BACKEND_URL;
 function ServicesContainer() {
   const servicesData = useSelector((state) => state.servicesReducer.servicesData);
   const dispatch = useDispatch();
-
-  useEffect(() => {
-    (async () => {
-      let res = await Axios.get(`${backendUrl}/services`);
-    
-      dispatch({type: 'UPDATE_SERVICES_DATA', newServicesData: res.data})
-    })();
-  }, []);
   
   return (
     <section className="ServicesContainer">
